readding the old splitter
