+++ prover/example/test_cases/framework_test.go
@@ -17,7 +17,6 @@ import (
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/permutation"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/specialqueries"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter/sticker"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/univariates"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/vortex"
 	"github.com/consensys/linea-monorepo/prover/protocol/ifaces"
@@ -70,7 +69,6 @@ var (
 		innerproduct.Compile,
 	}
 	ARITHMETICS = compilationSuite{
-		sticker.Sticker(4, 8),
 		splitter.SplitColumns(8),
 		localcs.Compile,
 		globalcs.Compile,
