+++ prover/example/test_cases/local_opening_test.go
@@ -12,7 +12,6 @@ import (
 	"github.com/consensys/linea-monorepo/prover/maths/field"
 	"github.com/consensys/linea-monorepo/prover/protocol/column"
 	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter"
-	"github.com/consensys/linea-monorepo/prover/protocol/compiler/splitter/sticker"
 	"github.com/consensys/linea-monorepo/prover/protocol/wizard"
 	"github.com/stretchr/testify/require"
 )
@@ -32,7 +31,7 @@ func proverLocalOpening(run *wizard.ProverRuntime) {
 }
 
 func TestGnarkCompile(t *testing.T) {
-	comp := wizard.Compile(defineLocalOpening, sticker.Sticker(8, 32), splitter.SplitColumns(32))
+	comp := wizard.Compile(defineLocalOpening, splitter.SplitColumns(32))
 	proof := wizard.Prove(comp, proverLocalOpening)
 
 	circ, err := wizard.AllocateWizardCircuit(comp)
